+++ stingray/__init__.py
@@ -22,3 +22,5 @@ if not _ASTROPY_SETUP_:
     from stingray.stats import *
     from stingray.bispectrum import *
     from stingray.varenergyspectrum import *
+    from stingray.largememory import *
+    from stingray.lombscargle import *
